+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Filter illegal chars in ingredient search
 - Added orientation when exporting to a formula for selling
 - Allow ordering in formula sell table
+- Set custom decimal precission when generating a formula to sell
 ### Version 10.0
 - Discord server link added
 - Added archive option when deleting a formula