+++ releasenotes.md
@@ -1,3 +1,7 @@
 Whats New in v10.1
 --------------------------
+- Set custom decimal precission when generating a formula to sell
+- Added orientation when exporting to a formula for selling
+- Allow ordering in formula sell table
+
 - For full details please refer to the CHANGELOG